YT-2130 - Implement fractional CPU resources
- Add FixedPointNumber class to store CPU usage efficiently
- Replace cpu type in scheduler from int to TCpuResource
- Replace cpu type in job_proxy from int to double

+++ yt/server/scheduler/fair_share_tree.h
@@ -21,7 +21,7 @@ struct TSchedulableAttributes
     double MaxPossibleUsageRatio = 1.0;
     double BestAllocationRatio = 1.0;
     double GuaranteedResourcesRatio = 0.0;
-    i64 DominantLimit = 0;
+    double DominantLimit = 0;
     int FifoIndex = 0;
 
     double AdjustedFairShareStarvationTolerance = 1.0;
